+++ src/main/java/fi/solita/jsonmigraine/internal/DataVersion.java
@@ -1,4 +1,4 @@
-// Copyright © 2012 Solita Oy <www.solita.fi>
+// Copyright © 2012-2013 Solita Oy <www.solita.fi>
 // This software is released under the MIT License.
 // The license text is at http://opensource.org/licenses/MIT
 
@@ -6,10 +6,14 @@ package fi.solita.jsonmigraine.internal;
 
 public class DataVersion {
 
-    public final Class<?> dataType;
+    public final String dataType;
     public final int dataVersion;
 
     public DataVersion(Class<?> dataType, int dataVersion) {
+        this(dataType.getName(), dataVersion);
+    }
+
+    public DataVersion(String dataType, int dataVersion) {
         this.dataType = dataType;
         this.dataVersion = dataVersion;
     }